# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ApexFinder Pro - Enterprise OSINT & HNWI Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /research/sessions/{id}/pitch:
    post:
      operationId: generatePitch
      tags: [research]
      summary: Generate an AI-personalized outreach pitch for the winning MCTS path
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchSession"
        "404":
          description: Not found

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ResearchSession:
      type: object
      required: [id, targetEntityId, crmStatus, createdAt]
      properties:
        id: { type: integer }
        targetEntityId: { type: integer }
        targetEntityName: { type: ["string", "null"] }
        winningPath: { type: ["string", "null"] }
        mctsSteps: { type: ["string", "null"] }
        generatedPitch: { type: ["string", "null"] }
        crmStatus: { type: string }
        lastContactDate: { type: ["string", "null"] }
        bayesianScoreAtRuntime: { type: ["number", "null"] }
        pathScore: { type: ["number", "null"] }
        realismClass: { type: ["string", "null"] }
        realismReason: { type: ["string", "null"] }
        researchBudget: { type: ["string", "null"] }
        createdAt: { type: string }
...